backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pymongo>=4.8.0
python-dotenv>=1.0.1
requests>=2.32.0
spacy>=3.7.5